freud-kb/requirements.txt
beautifulsoup4
lxml
tqdm
requests
pdfplumber
chromadb
python-dotenv
python-docx
gradio